scripts/wta_projects/auto_triage/requirements.txt
# Flask Framework
Flask
Werkzeug
click
itsdangerous
Jinja2
MarkupSafe
python-dotenv

# Data Processing
pandas
numpy

# Database
mysql-connector-python
pymysql
sqlalchemy

# JIRA Integration
jira
atlassian-python-api

# Image Processing
Pillow
opencv-python
scikit-image

# Utilities
python-dateutil
pytz==2023.3
requests==2.31.0

# Similarity Comparison
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
sentence-transformers

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0

# Additional dependencies
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0
